chat/requirements.txt
openai
httpx
python-dotenv
mcp